Added timer and button
The timer runs on a script that calculates the delta time passed between
frames and subtracts it from the total time in seconds. That time is
then mathematically converted into hours, minutes, and seconds and
displayed to the user.

A digital font I found online was used for the clock's text. It's
licensed as free for personal and educational use. I had to generate
a TextMeshPro font asset to use it. It appears to work fine.

A button has been placed in the scen, but it isn't interactable in VR
yet.